+++ src/main/kotlin/com/s63d/ervehicleservice/clients/AccountClient.kt
@@ -1,6 +1,6 @@
 package com.s63d.ervehicleservice.clients
 
-import com.s63d.ervehicleservice.domain.SimpleAccount
+import com.s63d.ervehicleservice.domain.db.SimpleAccount
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
