+++ src/main/kotlin/com/s63d/ervehicleservice/config/SecurityConfig.kt
@@ -20,6 +20,8 @@ class SecurityConfig(private val userDetailsService: UserDetailsService) : WebSe
         http.cors()
                 .and().csrf().disable()
                 .authorizeRequests()
+                .antMatchers("/v2/api-docs","/swagger-resources","/swagger-resources/**", "/swagger-ui.html", "/webjars/**").permitAll()
+                .antMatchers("/actuator/**").permitAll()
                 .anyRequest().authenticated()
                 .and()
                 .addFilterAfter(JwtAuthorizationFilter(authenticationManager(), userDetailsService), BasicAuthenticationFilter::class.java)
