+++ src/main/kotlin/com/s63d/ervehicleservice/repositories/AccountRepository.kt
@@ -1,6 +1,6 @@
 package com.s63d.ervehicleservice.repositories
 
-import com.s63d.ervehicleservice.domain.SimpleAccount
+import com.s63d.ervehicleservice.domain.db.SimpleAccount
 import org.springframework.data.repository.CrudRepository
 
 interface AccountRepository : CrudRepository<SimpleAccount, Long>
