+++ src/main/kotlin/com/s63d/ervehicleservice/repositories/CarTrackerRepository.kt
@@ -1,6 +1,6 @@
 package com.s63d.ervehicleservice.repositories
 
-import com.s63d.ervehicleservice.domain.CarTracker
+import com.s63d.ervehicleservice.domain.db.CarTracker
 import org.springframework.data.repository.CrudRepository
 
 interface CarTrackerRepository : CrudRepository<CarTracker, String>
