+++ src/main/kotlin/com/s63d/ervehicleservice/repositories/VehicleRepository.kt
@@ -1,6 +1,6 @@
 package com.s63d.ervehicleservice.repositories
 
-import com.s63d.ervehicleservice.domain.Vehicle
+import com.s63d.ervehicleservice.domain.db.Vehicle
 import org.springframework.data.repository.CrudRepository
 
 interface VehicleRepository : CrudRepository<Vehicle, String> {
