+++ src/main/kotlin/com/s63d/ervehicleservice/services/AccountService.kt
@@ -1,6 +1,6 @@
 package com.s63d.ervehicleservice.services
 
-import com.s63d.ervehicleservice.domain.SimpleAccount
+import com.s63d.ervehicleservice.domain.db.SimpleAccount
 import com.s63d.ervehicleservice.repositories.AccountRepository
 import org.springframework.stereotype.Service
