+++ src/main/kotlin/com/s63d/ervehicleservice/services/OwnershipService.kt
@@ -1,17 +1,22 @@
 package com.s63d.ervehicleservice.services
 
-import com.s63d.ervehicleservice.domain.Ownership
-import com.s63d.ervehicleservice.domain.SimpleAccount
-import com.s63d.ervehicleservice.domain.Vehicle
+import com.s63d.ervehicleservice.domain.db.Ownership
+import com.s63d.ervehicleservice.domain.db.SimpleAccount
+import com.s63d.ervehicleservice.domain.db.Vehicle
+import com.s63d.ervehicleservice.repositories.AccountRepository
 import com.s63d.ervehicleservice.repositories.OwnershipRepository
 import org.springframework.stereotype.Service
 import java.util.*
 
 @Service
-class OwnershipService(private val ownershipRepository: OwnershipRepository) {
+class OwnershipService(private val ownershipRepository: OwnershipRepository, private val accountRepository: AccountRepository) {
 
     fun getAllForUser(account: SimpleAccount) = ownershipRepository.getByAccount(account)
 
+    fun getAllForUser(accountId: String): List<Ownership> {
+        val account = accountRepository.findById(accountId.toLong()).orElseThrow { Exception("Account id is not found") } // TODO
+        return getAllForUser(account)
+    }
     fun createNew(vehicle: Vehicle, account: SimpleAccount) : Ownership = ownershipRepository.save(Ownership(account = account, vehicle = vehicle))
 
     fun suspend(vehicle: Vehicle, account: SimpleAccount) : Ownership {
