+++ src/main/kotlin/com/s63d/ervehicleservice/services/VehicleService.kt
@@ -1,6 +1,7 @@
 package com.s63d.ervehicleservice.services
 
-import com.s63d.ervehicleservice.domain.Vehicle
+import com.s63d.ervehicleservice.domain.db.Ownership
+import com.s63d.ervehicleservice.domain.db.Vehicle
 import com.s63d.ervehicleservice.repositories.VehicleRepository
 import com.s63d.ervehicleservice.utils.md5
 import org.springframework.stereotype.Service
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Service
 @Service
 class VehicleService(internal val vehicleRepository: VehicleRepository, private val ownershipService: OwnershipService, private val accountService: AccountService) {
 
-    fun createNew(license: String, type: String, brand: String, color: String, weight: Int, accountId: Long): Vehicle {
+    fun createNew(license: String, type: String, brand: String, color: String, weight: Int, accountId: Long): Ownership {
         val account = accountService.getOrCreate(accountId)
         val vehicle = Vehicle(type, brand, color, weight, 'A', license.md5())
 
@@ -16,8 +17,7 @@ class VehicleService(internal val vehicleRepository: VehicleRepository, private 
             throw Exception("License already registered")
 
         vehicleRepository.save(vehicle)
-        ownershipService.createNew(vehicle, account)
-        return vehicle
+        return ownershipService.createNew(vehicle, account)
     }
 
     fun getForAccount(accountId: Long) = accountService.findById(accountId).ownerships.map { it.vehicle }
